# swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is db course project backend API.",
        "title": "DB course project API",
        "contact": {
            "name": "API Support",
            "email": "evgeniazavojskih@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/analyzer": {
            "post": {
                "description": "add new analyzer",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Add new analyzer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Information about stored analyzer",
                        "name": "analyzer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db_cp_6_sem_internal_domain_entity.Analyzer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analyzer/{id}": {
            "get": {
                "description": "return analyzer with the specified id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Show analyzer with the specified id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Analyzer id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "delete analyzer by id",
                "tags": [
                    "employee"
                ],
                "summary": "Delete analyzer by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Analyzer id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/analyzer_type": {
            "post": {
                "description": "add new analyzer type",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Add new analyzer type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Information about stored analyzer type",
                        "name": "anType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db_cp_6_sem_internal_domain_entity.CreateType"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analyzer_type/{name}": {
            "delete": {
                "description": "delete analyzer type by name",
                "tags": [
                    "employee"
                ],
                "summary": "Delete analyzer type by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Analyzer type name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/event": {
            "post": {
                "description": "add new event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Add new event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Information about stored event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db_cp_6_sem_internal_domain_entity.CreateEvent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/event/{id}": {
            "delete": {
                "description": "delete event by id",
                "tags": [
                    "employee"
                ],
                "summary": "Delete event by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/gas": {
            "post": {
                "description": "add new gas",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Add new gas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Information about stored gas",
                        "name": "gas",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db_cp_6_sem_internal_domain_entity.Gas"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/sensor": {
            "post": {
                "description": "add new sensor",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Add new sensor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Information about stored sensor",
                        "name": "sensor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db_cp_6_sem_internal_domain_entity.Sensor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sensor/{id}": {
            "delete": {
                "description": "delete sensor by id",
                "tags": [
                    "employee"
                ],
                "summary": "Delete sensor by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sensor id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "description": "update sensor analyzer_id by id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Update sensor analyzer_id by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "New analyzer id",
                        "name": "anId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Sensor id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db_cp_6_sem_internal_domain_entity.Analyzer": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "job_number": {
                    "type": "string"
                },
                "part_number": {
                    "type": "string"
                },
                "software_version": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "db_cp_6_sem_internal_domain_entity.CreateEvent": {
            "type": "object",
            "properties": {
                "peak_readings": {
                    "type": "number"
                },
                "sensor_id": {
                    "type": "string"
                },
                "signal_time": {
                    "type": "string"
                }
            }
        },
        "db_cp_6_sem_internal_domain_entity.CreateType": {
            "type": "object",
            "properties": {
                "gases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "max_sensors": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "db_cp_6_sem_internal_domain_entity.Gas": {
            "type": "object",
            "properties": {
                "formula": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "db_cp_6_sem_internal_domain_entity.Sensor": {
            "type": "object",
            "properties": {
                "analyzer_id": {
                    "type": "string"
                },
                "gas": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "low_limit_alarm": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "upper_limit_alarm": {
                    "type": "string"
                }
            }
        }
    }
}